fixup! Move deploy method to manager
